vim: Add flags for plugin configs for different environments
The SIMPLE_VIM_PLUGINS environmental variable turns off plugins with a
usage cost (e.g. airline) or complicate the config (e.g. snippets).

The STANDALONE_VIM_PLUGINS environmental variable turns off plugins that
use external programs and aren't pure vimL.
This is useful for creating a config that will be copied to a minimal
environment such as a phone.